reworked core components. need to reimplement partial import
